dvc-data: move load/check/iterobjs to hashfile
Leftover from #164
